+++ internal_filesystem/lib/mpos/app/__init__.py
@@ -1,5 +1,6 @@
 from .app import App
 from .activity import Activity
+from .service import Service
 from .activities.chooser import ChooserActivity
 from .activities.view import ViewActivity
 from .activities.share import ShareActivity